+++ protocol-code-generator/src/main/java/dev/cirras/generate/ObjectCodeGenerator.java
@@ -111,6 +111,7 @@ final class ObjectCodeGenerator {
                     .orElse(null))
             .arrayField(true)
             .delimited(protocolArray.isDelimited())
+            .trailingDelimiter(protocolArray.hasTrailingDelimiter())
             .build();
 
     fieldCodeGenerator.generateField();