+++ protocol-code-generator/src/main/java/dev/cirras/xml/ProtocolArray.java
@@ -1,5 +1,6 @@
 package dev.cirras.xml;
 
+import jakarta.xml.bind.Unmarshaller;
 import jakarta.xml.bind.annotation.XmlAttribute;
 import jakarta.xml.bind.annotation.XmlElement;
 import jakarta.xml.bind.annotation.XmlRootElement;
@@ -18,6 +19,9 @@ public class ProtocolArray {
   @XmlAttribute private boolean optional;
 
   @XmlAttribute private boolean delimited;
+
+  @XmlAttribute(name = "trailing-delimiter")
+  private Boolean trailingDelimiter;
 
   @XmlElement private ProtocolComment comment;
 
@@ -41,7 +45,18 @@ public class ProtocolArray {
     return delimited;
   }
 
+  public boolean hasTrailingDelimiter() {
+    return trailingDelimiter;
+  }
+
   public Optional<ProtocolComment> getComment() {
     return Optional.ofNullable(comment);
+  }
+
+  @SuppressWarnings("unused")
+  private void afterUnmarshal(Unmarshaller unmarshaller, Object parent) {
+    if (trailingDelimiter == null) {
+      trailingDelimiter = delimited;
+    }
   }
 }